requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.1.6
django-bootstrap-form==3.4
django-heroku==0.3.1
django-role-permissions==3.2.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pypdf==5.1.0
PyPDF2==3.0.1
python-decouple==3.8
qrcode==8.0
reportlab==4.2.5
setuptools==75.6.0
sqlparse==0.5.3
static3==0.7.0
types-PyYAML==6.0.12.20241221
typing_extensions==4.12.2
whitenoise==6.9.0